cine-api/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
networkx
matplotlib
motor
python-dotenv